moved forward calculations into forward folder
