+++ app/lib/calculations/forwards/commons_only_sitting_days.rb
